fix: SugiyamaLayoutEngine と SequenceDiagramRenderer のコンパイルエラー修正
- double→float の明示的キャスト追加
- using UnityEngine.UIElements 追加
- Position.Absolute の完全修飾名使用

+++ unity/Packages/com.mermaid-graphview/Editor/Layout/SugiyamaLayoutEngine.cs
@@ -317,7 +317,7 @@ namespace MermaidGraphView
                 }
 
                 if (connectedPositions.Count > 0)
-                    barycenters[nodeId] = connectedPositions.Average();
+                    barycenters[nodeId] = (float)connectedPositions.Average();
                 else
                     barycenters[nodeId] = positions.ContainsKey(nodeId) ? positions[nodeId] : 0;
             }
